Migration: Separate legacy and index entrypoint (#54124)
