+++ src/legacy/core_plugins/kibana/index.js
@@ -60,7 +60,12 @@ export default function(kibana) {
     },
 
     uiExports: {
-      hacks: ['plugins/kibana/discover', 'plugins/kibana/dev_tools', 'plugins/kibana/visualize'],
+      hacks: [
+        'plugins/kibana/discover/legacy',
+        'plugins/kibana/dev_tools',
+        'plugins/kibana/visualize/legacy',
+        'plugins/kibana/dashboard/legacy',
+      ],
       savedObjectTypes: ['plugins/kibana/dashboard/saved_dashboard/saved_dashboard_register'],
       app: {
         id: 'kibana',
