+++ src/legacy/core_plugins/kibana/public/discover/__tests__/query_parameters/action_add_filter.js
@@ -22,7 +22,7 @@ import ngMock from 'ng_mock';
 import { createStateStub } from './_utils';
 import { getQueryParameterActions } from '../../np_ready/angular/context/query_parameters/actions';
 import { createIndexPatternsStub } from '../../np_ready/angular/context/api/__tests__/_stubs';
-import { pluginInstance } from 'plugins/kibana/discover/index';
+import { pluginInstance } from 'plugins/kibana/discover/legacy';
 import { npStart } from 'ui/new_platform';
 
 describe('context app', function() {
