+++ src/legacy/core_plugins/kibana/public/discover/index.ts
@@ -16,24 +16,13 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-import chrome from 'ui/chrome';
+
 import { PluginInitializerContext } from 'kibana/public';
-import { npSetup, npStart } from 'ui/new_platform';
 import { DiscoverPlugin } from './plugin';
+
+export { createSavedSearchesService } from './saved_searches/saved_searches';
 
 // Core will be looking for this when loading our plugin in the new platform
 export const plugin = (context: PluginInitializerContext) => {
   return new DiscoverPlugin();
 };
-
-// Legacy compatiblity part - to be removed at cutover, replaced by a kibana.json file
-export const pluginInstance = plugin({} as PluginInitializerContext);
-export const setup = pluginInstance.setup(npSetup.core, {
-  ...npSetup.plugins,
-  __LEGACY: {
-    chrome,
-  },
-});
-export const start = pluginInstance.start(npStart.core, npStart.plugins);
-
-export { createSavedSearchesService } from './saved_searches/saved_searches';
