+++ src/legacy/core_plugins/kibana/public/discover/legacy.ts
@@ -16,13 +16,17 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-
+import chrome from 'ui/chrome';
 import { PluginInitializerContext } from 'kibana/public';
-import { DashboardPlugin } from './plugin';
+import { npSetup, npStart } from 'ui/new_platform';
+import { plugin } from './index';
 
-export * from './np_ready/dashboard_constants';
-
-// Core will be looking for this when loading our plugin in the new platform
-export const plugin = (context: PluginInitializerContext) => {
-  return new DashboardPlugin();
-};
+// Legacy compatiblity part - to be removed at cutover, replaced by a kibana.json file
+export const pluginInstance = plugin({} as PluginInitializerContext);
+export const setup = pluginInstance.setup(npSetup.core, {
+  ...npSetup.plugins,
+  __LEGACY: {
+    chrome,
+  },
+});
+export const start = pluginInstance.start(npStart.core, npStart.plugins);
