+++ src/legacy/core_plugins/kibana/public/discover/np_ready/angular/context/api/__tests__/anchor.js
@@ -19,7 +19,7 @@
 
 import expect from '@kbn/expect';
 import ngMock from 'ng_mock';
-import { pluginInstance } from 'plugins/kibana/discover/index';
+import { pluginInstance } from 'plugins/kibana/discover/legacy';
 
 import { createIndexPatternsStub, createSearchSourceStub } from './_stubs';
