+++ src/legacy/core_plugins/kibana/public/home/index.ts
@@ -22,10 +22,7 @@ import { npSetup, npStart } from 'ui/new_platform';
 import chrome from 'ui/chrome';
 import { IPrivate } from 'ui/private';
 import { HomePlugin, LegacyAngularInjectedDependencies } from './plugin';
-import { createUiStatsReporter, METRIC_TYPE } from '../../../ui_metric/public';
 import { TelemetryOptInProvider } from '../../../telemetry/public/services';
-
-export const trackUiMetric = createUiStatsReporter('Kibana_home');
 
 /**
  * Get dependencies relying on the global angular context.
@@ -54,9 +51,7 @@ let copiedLegacyCatalogue = false;
   instance.setup(npSetup.core, {
     ...npSetup.plugins,
     __LEGACY: {
-      trackUiMetric,
       metadata: npStart.core.injectedMetadata.getLegacyMetadata(),
-      METRIC_TYPE,
       getFeatureCatalogueEntries: async () => {
         if (!copiedLegacyCatalogue) {
           const injector = await chrome.dangerouslyGetActiveInjector();
