+++ src/legacy/core_plugins/kibana/public/kibana.js
@@ -47,9 +47,9 @@ import 'uiExports/interpreter';
 import 'ui/autoload/all';
 import 'ui/kbn_top_nav';
 import './home';
-import './discover';
-import './visualize';
-import './dashboard';
+import './discover/legacy';
+import './visualize/legacy';
+import './dashboard/legacy';
 import './management';
 import './dev_tools';
 import 'ui/color_maps';
