+++ x-pack/legacy/plugins/dashboard_mode/public/dashboard_viewer.js
@@ -34,6 +34,7 @@ import 'ui/color_maps';
 import 'ui/agg_response';
 import 'ui/agg_types';
 import 'leaflet';
+import 'plugins/kibana/dashboard/legacy';
 import { npStart } from 'ui/new_platform';
 import { localApplicationService } from 'plugins/kibana/local_application_service';
